+++ railties/test/application/integration_test_case_test.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
 require "isolation/abstract_unit"
+require "env_helpers"
 
 module ApplicationTests
   class IntegrationTestCaseTest < ActiveSupport::TestCase
-    include ActiveSupport::Testing::Isolation
+    include ActiveSupport::Testing::Isolation, EnvHelpers
 
     setup do
       build_app
@@ -39,13 +40,14 @@ module ApplicationTests
         end
       RUBY
 
+      with_rails_env("test") { rails("db:migrate") }
       output = rails("test")
       assert_match(/0 failures, 0 errors/, output)
     end
   end
 
   class IntegrationTestDefaultApp < ActiveSupport::TestCase
-    include ActiveSupport::Testing::Isolation
+    include ActiveSupport::Testing::Isolation, EnvHelpers
 
     setup do
       build_app
@@ -66,6 +68,7 @@ module ApplicationTests
         end
       RUBY
 
+      with_rails_env("test") { rails("db:migrate") }
       output = rails("test")
       assert_match(/0 failures, 0 errors/, output)
     end